+++ srcs/mods/ft_ll_mod.c
@@ -6,7 +6,7 @@
 /*   By: jagarcia <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/14 01:33:10 by jagarcia          #+#    #+#             */
-/*   Updated: 2018/01/29 22:36:05 by jagarcia         ###   ########.fr       */
+/*   Updated: 2018/01/30 11:05:31 by jagarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ char	*ft_ll_mod(va_list ap, va_list ap2, char *comm)
 	
 	comand = comm[ft_strlen(comm) - 1];
 	if ((comand == 'd') || (comand == 'i'))
-		return (ft_ltoa_base(*((long *)ft_locate_date(comm,
+		return (ft_ltoa_base(*((long long *)ft_locate_date(comm,
 							10, ap, ap2)), 10));
 	if (comand == 'u')
-		return (ft_ultoa_base(*((unsigned long *)ft_locate_date(comm,
+		return (ft_ultoa_base(*((unsigned long long *)ft_locate_date(comm,
 							11, ap, ap2)), 10));
 	if (comand == 'o')
 		return (ft_dectooct(ft_locate_date(comm, 11, ap, ap2),
