+++ srcs/mods/ft_z_mod.c
@@ -6,7 +6,7 @@
 /*   By: jagarcia <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/14 01:33:10 by jagarcia          #+#    #+#             */
-/*   Updated: 2018/01/29 22:34:15 by jagarcia         ###   ########.fr       */
+/*   Updated: 2018/01/30 11:07:22 by jagarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ char	*ft_z_mod(va_list ap, va_list ap2, char *comm)
 
 	comand = comm[ft_strlen(comm) - 1];
 	if ((comand == 'd') || (comand == 'i'))
-		return (ft_ltoa_base(*((long *)ft_locate_date(comm,
+		return (ft_ltoa_base(*((size_t *)ft_locate_date(comm,
 							14, ap, ap2)), 10));
 	if (comand == 'u')
-		return (ft_ultoa_base(*((unsigned long *)ft_locate_date(comm,
+		return (ft_ultoa_base(*((size_t *)ft_locate_date(comm,
 							14, ap, ap2)), 10));
 	if (comand == 'o')
 		return (ft_dectooct(ft_locate_date(comm, 14, ap, ap2),
